Load artist albums using artist id/fixed album cover not shown in artist details
Album art issue related to #48, and is fixed by using long id's.

+++ app/src/main/java/com/hardcodecoder/pulsemusic/model/ArtistModel.java
@@ -4,19 +4,19 @@ import androidx.annotation.NonNull;
 
 public class ArtistModel {
 
-    private final int mArtistId;
+    private final long mArtistId;
     private final int mNumOfAlbums;
     private final int mNumOfTracks;
     private final String mArtistName;
 
-    public ArtistModel(int artistId, @NonNull String artistName, int numOfAlbums, int numOfTracks) {
+    public ArtistModel(long artistId, @NonNull String artistName, int numOfAlbums, int numOfTracks) {
         mArtistId = artistId;
         mNumOfAlbums = numOfAlbums;
         mNumOfTracks = numOfTracks;
         mArtistName = artistName;
     }
 
-    public int getArtistId() {
+    public long getArtistId() {
         return mArtistId;
     }
